Donor Stuff | Adds support for custom azure crowns + things for morto (again)

+++ code/modules/clothing/rogueclothes/crown.dm
@@ -18,15 +18,20 @@
 	var/messagereceivedsound = 'sound/misc/scom.ogg'
 	var/hearrange = 0 // Only hearable by wearer
 	is_important = TRUE
+	var/register_as_roguemachine_crown = TRUE // New vars to allow custom donor crowns. Does not prevent custom crowns from being replaced by the original, though.
+	var/replace_existing_roguemachine_crown = FALSE
 
 /obj/item/clothing/head/roguetown/crown/serpcrown/get_examine_highlight_status()
 	return list(EXAMINEHIGHLIGHT_VIBE_CROWN, VIBEDESC_CROWN)
 
 /obj/item/clothing/head/roguetown/crown/serpcrown/Initialize(mapload)
 	. = ..()
-	if(SSroguemachine.crown)
-		qdel(src)
-	else
+	if(register_as_roguemachine_crown)
+		if(SSroguemachine.crown)
+			if(!replace_existing_roguemachine_crown)
+				qdel(src)
+				return
+			qdel(SSroguemachine.crown)
 		SSroguemachine.crown = src
 		SSroguemachine.scomm_machines += src
 	become_hearing_sensitive()
@@ -121,6 +126,9 @@
 		send_speech(message, hearrange, src, , spans, message_language=language)
 
 /obj/item/clothing/head/roguetown/crown/serpcrown/Destroy()
+	SSroguemachine.scomm_machines -= src
+	if(SSroguemachine.crown == src)
+		SSroguemachine.crown = null
 	lose_hearing_sensitivity()
 	return ..()
